say: Return text as a map
This is the first phase of moving text rendering for log messages to
the client. The :text field now returns a map, which contains
the username (depending on the action) and raw-text, which contains
the text without the default "<username> <text>." rendering.

The client side in this change now takes care of the rendering of the
username (if present) and raw text.

Text generation will incrementally move into additional fields in this
map, leading to an incremental process where more and more logic can
be moved to the client.